// src/generated/graphql.tsx
export type Maybe<T> = T | null;
/** All built-in and custom scalars, mapped to their actual values */
export type Scalars = {
  ID: string;
  String: string;
  Boolean: boolean;
  Int: number;
  Float: number;
  /** An RFC 3986, RFC 3987, and RFC 6570 (level 4) compliant URI string. */
  URI: any;
  /** An ISO-8601 encoded UTC date string. */
  DateTime: any;
  /** A string containing HTML code. */
  HTML: any;
  /** A Git object ID. */
  GitObjectID: any;
  /** An ISO-8601 encoded date string. Unlike the DateTime type, GitTimestamp is not converted in UTC. */
  GitTimestamp: any;
  /** Git SSH string */
  GitSSHRemote: any;
  /** An ISO-8601 encoded date string. */
  Date: any;
  /** An ISO-8601 encoded UTC date string with millisecond precison. */
  PreciseDateTime: any;
  /** A valid x509 certificate string */
  X509Certificate: any;
};
// ...
export type Blob = GitObject &
  Node & {
    __typename?: "Blob";
    /** An abbreviated version of the Git object ID */
    abbreviatedOid: Scalars["String"];
    /** Byte size of Blob object */
    byteSize: Scalars["Int"];
    /** The HTTP path for this Git object */
    commitResourcePath: Scalars["URI"];
    /** The HTTP URL for this Git object */
    commitUrl: Scalars["URI"];
    id: Scalars["ID"];
    /** Indicates whether the Blob is binary or text */
    isBinary: Scalars["Boolean"];
    /** Indicates whether the contents is truncated */
    isTruncated: Scalars["Boolean"];
    /** The Git object ID */
    oid: Scalars["GitObjectID"];
    /** The Repository the Git object belongs to */
    repository: Repository;
    /** UTF8 text data or null if the Blob is binary */
    text?: Maybe<Scalars["String"]>;
  };
// ...
export type GitObject = {
  __typename?: "GitObject";
  /** An abbreviated version of the Git object ID */
  abbreviatedOid: Scalars["String"];
  /** The HTTP path for this Git object */
  commitResourcePath: Scalars["URI"];
  /** The HTTP URL for this Git object */
  commitUrl: Scalars["URI"];
  id: Scalars["ID"];
  /** The Git object ID */
  oid: Scalars["GitObjectID"];
  /** The Repository the Git object belongs to */
  repository: Repository;
};
// ...
export type Node = {
  __typename?: "Node";
  /** ID of the object. */
  id: Scalars["ID"];
};